26/05: fix acc co khu, phong co thiet bi bao khong duoc xoa va khong …

+++ src/java/controller/DeleteAcc.java
@@ -2,10 +2,10 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
  */
-
 package controller;
 
 import dal.DAO;
+import jakarta.servlet.RequestDispatcher;
 import java.io.IOException;
 import java.io.PrintWriter;
 import jakarta.servlet.ServletException;
@@ -21,32 +21,48 @@ import java.util.logging.Logger;
  *
  * @author admin
  */
-@WebServlet(name="DeleteAcc", urlPatterns={"/deleteacc"})
+@WebServlet(name = "DeleteAcc", urlPatterns = {"/deleteacc"})
 public class DeleteAcc extends HttpServlet {
-   
-    /** 
-     * Processes requests for both HTTP <code>GET</code> and <code>POST</code> methods.
+
+    /**
+     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
+     * methods.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
      * @throws IOException if an I/O error occurs
      */
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
         String aid = request.getParameter("aid");
         DAO dao = new DAO();
-        try {
-            dao.deleteAccount(aid);
-        } catch (SQLException ex) {
-            Logger.getLogger(DeleteAcc.class.getName()).log(Level.SEVERE, null, ex);
+        if (dao.checkAccIDcoKhu(aid)) {
+            request.setAttribute("error", "Account đang quản lý khu không thể xóa!");
+            RequestDispatcher dispatcher = request.getRequestDispatcher("listaccount");
+            dispatcher.forward(request, response);
+        } else {
+            try {
+                if ("1".equals(aid)) {
+                    request.setAttribute("error", "Không Thể Xóa Account Admin");
+                    RequestDispatcher dispatcher = request.getRequestDispatcher("listaccount");
+                    dispatcher.forward(request, response);
+                    return;
+                } else {
+                    dao.deleteAccount(aid);
+                }
+            } catch (SQLException ex) {
+                Logger.getLogger(DeleteAcc.class.getName()).log(Level.SEVERE, null, ex);
+            }
         }
         response.sendRedirect("listaccount");
-    } 
+    }
 
     // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-    /** 
+    /**
      * Handles the HTTP <code>GET</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -54,12 +70,13 @@ public class DeleteAcc extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
-    } 
+    }
 
-    /** 
+    /**
      * Handles the HTTP <code>POST</code> method.
+     *
      * @param request servlet request
      * @param response servlet response
      * @throws ServletException if a servlet-specific error occurs
@@ -67,12 +84,13 @@ public class DeleteAcc extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-    throws ServletException, IOException {
+            throws ServletException, IOException {
         processRequest(request, response);
     }
 
-    /** 
+    /**
      * Returns a short description of the servlet.
+     *
      * @return a String containing servlet description
      */
     @Override
